pyproject.toml
[project]
name = "fk2-hid-midi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["pynput", "mido", "python-rtmidi", "pywebview"]

[project.scripts]
fk2-hid-midi = "src.main:main"
